backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
